ATST-919 (часть 1)

See merge request qa/automag!1114

+++ web-tests/jdbc/src/main/java/ru/instamart/jdbc/entity/OffersEntity.java
@@ -31,5 +31,5 @@ public final class OffersEntity {
     private String quantityType;
     private Integer stock;
     private Integer maxStock;
-    private Byte published;
+    private Integer published;
 }
